+++ theme/src/typescript/admin/blocks/index.ts
@@ -1,6 +1,8 @@
+import opengraph from './opengraph';
 import speaker from './speaker';
 
 (() => {
   // Enabling on load.
+  opengraph.register();
   speaker.register();
 })();