+++ theme/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+- Feature: Added favicons
+- Feature: Created Open Graph metadata block
+
 ## 0.15.0
 
 - Feature: Added url field on supporters post